STYLE) IceCreamPainter 파일 이름 Camel 표기법에 맞도록 수정 및 paintIceCream 함수 리팩터링
